requirements.txt
matplotlib
networkx
dgl-cu113
torch
numpy
Pillow
pandas
nltk
scipy
scikit-learn
gensim
defusedxml
cffi
opencv-python
tqdm
requests
pydantic
PyYAML
psutil
pip
Jinja2
blis
h5py
click
regex
pyparsing
pytz
cymem
torchvision
pypdfium2
WeasyPrint
Unidecode
mplcursors
pyclipper
Shapely
rapidfuzz
setuptools
pathy
typer
spacy
thinc
srsly
wasabi
catalogue
packaging
langcodes
preshed
murmurhash
confection
six
urllib3
MarkupSafe
joblib
python-dateutil
certifi
threadpoolctl
cssselect2
tinycss2
cairocffi
webencodings
idna
pycparser
jarowinkler
cycler
kiwisolver
torchaudio
pyphen
CairoSVG
html5lib
filelock
ogb
outdated
littleutils
torchmetrics